auth in header

+++ src/main/java/com/example/auth/domain/post/post/controller/ApiV1PostController.java
@@ -40,7 +40,6 @@ public class ApiV1PostController {
         );
     }
 
-
     @GetMapping("{id}")
     public RsData<PostDto> getItem(@PathVariable long id) {
 
@@ -53,9 +52,26 @@ public class ApiV1PostController {
         );
     }
 
+    record DeleteReqBody(@NotNull Long authorId,
+                         @NotBlank @Length(min = 3) String password) {
+    }
+
     @DeleteMapping("/{id}")
-    public RsData<Void> delete(@PathVariable long id) {
+    public RsData<Void> delete(@PathVariable long id, @RequestHeader Long authorId,
+                               @RequestHeader String password) {
+
+        Member actor = memberService.findById(authorId).get();
+
+        if (!actor.getPassword().equals(password)) {
+            throw new ServiceException("401-1", "비밀번호가 일치하지 않습니다.");
+        }
+
         Post post = postService.getItem(id).get();
+
+        if(post.getAuthor().getId() != authorId) {
+            throw new ServiceException("403-1", "자신이 작성한 글만 삭제 가능합니다.");
+        }
+
         postService.delete(post);
 
         return new RsData<>(
@@ -63,7 +79,6 @@ public class ApiV1PostController {
                 "%d번 글 삭제가 완료되었습니다.".formatted(id)
         );
     }
-
 
     record ModifyReqBody(@NotBlank @Length(min = 3) String title,
                          @NotBlank @Length(min = 3) String content,
